meta-nexell: qtx11 image type append
This patch is about new image type 'qtx11'.
Qt5.4 application running with X11, based on core-image-sato.bb
See, recipes-core/images/nexell-qtx11.bb
Note that 'nexell-qt5.4' and 'qt-default-platform-x11' DISTRO_FEATURES
are added to distinguish qt5.4 from qt5.8.

+++ meta-nexell-distro/recipes-graphics/mesa/mesa_%.bbappend
@@ -1,7 +1,14 @@
-PROVIDES_remove = "virtual/libgles1 virtual/libgles2 virtual/libopenvg virtual/egl"
+PROVIDES_remove = " virtual/libgles1 virtual/libgles2 virtual/libopenvg "
+PROVIDES_remove += "${@bb.utils.contains('DISTRO_FEATURES', 'x11', '', 'virtual/egl', d)}"
+
+BACKEND = "${@bb.utils.contains('DISTRO_FEATURES', 'x11', 'x11', 'none', d)}"
 
 do_install_append () {
-    rm -f ${D}/${libdir}/libEGL*
+    if [ ${BACKEND} == "x11" ]; then
+       echo "nothing"
+    else
+       rm -f ${D}/${libdir}/libEGL*
+    fi
     rm -f ${D}/${libdir}/libGLESv1_CM.*
     rm -f ${D}/${libdir}/libGLESv2.*
     rm -f ${D}/${libdir}/libgbm*
